Introduce ChannelAuthorizer
